backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
monai>=1.3.0
nibabel>=5.1.0
pydicom>=2.4.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
matplotlib>=3.7.0
requests>=2.31.0
